requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
pyodbc==5.1.0

# Pydantic settings
pydantic-settings==2.1.0

# Claude API
anthropic==0.18.1

# Azure Speech (add later)
# azure-cognitiveservices-speech==1.35.0

# Development
python-dotenv==1.0.1
